Break apart FTPC
