+++ src/FTPC.jl
@@ -1,170 +1,6 @@
 using Compat: Cvoid, undef
 using LibCURL
 
-import Base: ==
-
-##############################
-# Type definitions
-##############################
-
-struct RequestOptions
-    uri::URI
-    ssl::Bool
-    verify_peer::Bool
-    active_mode::Bool
-end
-
-"""
-    RequestOptions(; kwargs...)
-
-The options used to connect to an FTP server.
-
-# Keywords
-- `hostname::AbstractString="localhost"`: the hostname or address of the FTP server.
-- `username::AbstractString=""`: the username used to access the FTP server.
-- `password::AbstractString=""`: the password used to access the FTP server.
-- `implicit::Bool=false`: use an implicit FTPS configuration.
-- `ssl::Bool=false`: use a secure connection. Typically specified for explicit FTPS.
-- `verify_peer::Bool=true`: verify authenticity of peer's certificate.
-- `active_mode::Bool=false`: use active mode to establish data connection.
-"""
-function RequestOptions(;
-    hostname::AbstractString="localhost",
-    port::Integer=0,
-    username::AbstractString="",
-    password::AbstractString="",
-    ssl::Bool=false,
-    implicit::Bool=false,
-    verify_peer::Bool=true,
-    active_mode::Bool=false,
-    url::AbstractString="",
-)
-    userinfo = if !isempty(password)
-        username * ":" * password
-    else
-        username
-    end
-
-    uri = if isempty(url)
-        scheme = ssl ? (implicit ? "ftps" : "ftpes") : "ftp"
-        URI(scheme, hostname, port, "", "", "", userinfo)
-    else
-        Base.depwarn(string(
-            "Using `RequestOptions` with the `url` keyword is deprecated; ",
-            "use `RequestOptions(url, ...)` instead",
-        ), :RequestOptions)
-        URI(URI(url); userinfo=userinfo)
-    end
-
-    RequestOptions(uri, ssl, verify_peer, active_mode)
-end
-
-function RequestOptions(
-    url::AbstractString;
-    verify_peer::Bool=true,
-    active_mode::Bool=false,
-    ssl::Union{Nothing, Bool}=nothing,
-)
-    if ssl !== nothing
-         Base.depwarn(
-            "`RequestOptions` keyword `ssl` has been depprecated, change the URL " *
-            "protocol to \"ftp://\", \"ftps://\", or \"ftpes://\" to respectively " *
-            "indicate no security, implicit security, or explicit security.",
-            :RequestOptions
-        )
-        url = ssl ? replace(url, "ftp://" => "ftpes://") : url
-    end
-
-    uri = URI(url)
-
-    if !(uri.scheme in ("ftps", "ftp", "ftpes"))
-        throw(ArgumentError("Unhandled FTP scheme: $(uri.scheme)"))
-    end
-
-    RequestOptions(
-        uri,
-        uri.scheme in ("ftps", "ftpes"),
-        verify_peer,
-        active_mode,
-    )
-end
-
-function security(opts::RequestOptions)
-    opts.ssl ? (opts.uri.scheme == "ftps" ? :implicit : :explicit) : :none
-end
-
-ispassive(opts::RequestOptions) = !opts.active_mode
-
-
-"""
-    Response
-
-The response returned from a connection to an FTP server.
-
-# Parameters
-* `body::IO`: contains the result of a command from ftp_command.
-    or the content of a downloaded file from ftp_get (if no destination file was defined).
-* `headers::Array{AbstractString}`: the header responses from the server.
-* `code::UInt`: the last header response code from the server.
-* `total_time::Float64`: the time the connection took.
-* `bytes_recd::Int`: the amount of bytes transmitted from the server (the file size in the
-    case of ftp_get).
-"""
-mutable struct Response
-    body::IO
-    headers::Array{AbstractString}
-    code::UInt
-    total_time::Float64
-    bytes_recd::Int
-
-    Response() = new(IOBuffer(), AbstractString[], 0, 0.0, 0)
-end
-
-function show(io::IO, o::Response)
-    println(io, "Response Code :", o.code)
-    println(io, "Request Time  :", o.total_time)
-    println(io, "Headers       :", o.headers)
-    println(io, "Length of body: ", o.bytes_recd)
-end
-
-mutable struct ReadData
-    src::IO
-    offset::Csize_t
-    sz::Csize_t
-
-    ReadData() = new(IOBuffer(), 0, 0)
-end
-
-mutable struct WriteData
-    buffer::IO
-    bytes_recd::Int
-
-    WriteData() = new(IOBuffer(), 0)
-end
-
-"""
-    ConnContext
-
-Keeps track of a persistent FTP connection.
-
-# Parameters
-* `url::AbstractString`: url of the FTP server.
-* `options::RequestOptions`: the options used for the connection.
-"""
-mutable struct ConnContext
-    curl::Ptr{CURL}
-    url::String  # Avoid using an abstract type when interacting with C libraries
-    options::RequestOptions
-
-    function ConnContext(options::RequestOptions)
-        new(C_NULL, trailing(string(options.uri), '/'), options)
-    end
-end
-
-function url(ctxt::ConnContext)
-    # Set ftpes protocol to ftp in the url so that libcurl can understand it
-    url = ctxt.options.ssl ? replace(ctxt.url, "ftpes://" => "ftp://") : ctxt.url
-end
 
 ##############################
 # Callbacks
@@ -229,62 +65,6 @@ macro ce_curl(f, args...)
     end
 end
 
-function setup_easy_handle(options::RequestOptions)
-    ctxt = ConnContext(options)
-
-    curl = curl_easy_init()
-    curl == C_NULL && error("curl_easy_init() failed")
-
-    ctxt.curl = curl
-
-    p_ctxt = pointer_from_objref(ctxt)
-
-    @ce_curl curl_easy_setopt CURLOPT_URL url(ctxt)
-    # @ce_curl curl_easy_setopt CURLOPT_VERBOSE Int64(1)
-
-    if options.ssl
-        @ce_curl curl_easy_setopt CURLOPT_USE_SSL CURLUSESSL_ALL
-        @ce_curl curl_easy_setopt CURLOPT_SSL_VERIFYHOST Int64(2)
-        @ce_curl curl_easy_setopt CURLOPT_FTPSSLAUTH CURLFTPAUTH_SSL
-        @ce_curl curl_easy_setopt CURLOPT_SSL_VERIFYPEER Int64(options.verify_peer)
-    end
-
-    if options.active_mode
-        @ce_curl curl_easy_setopt CURLOPT_FTPPORT "-"
-    end
-
-    return ctxt
-end
-
-function cleanup_easy_context(ctxt::ConnContext)
-    ctxt.curl == C_NULL && return nothing
-
-    # cleaning up should not write any data
-    @ce_curl curl_easy_setopt CURLOPT_WRITEFUNCTION C_NULL
-    @ce_curl curl_easy_setopt CURLOPT_WRITEDATA C_NULL
-
-    @ce_curl curl_easy_setopt CURLOPT_HEADERFUNCTION C_NULL
-    @ce_curl curl_easy_setopt CURLOPT_HEADERDATA C_NULL
-
-    @ce_curl curl_easy_setopt CURLOPT_READDATA C_NULL
-    @ce_curl curl_easy_setopt CURLOPT_READFUNCTION C_NULL
-
-    curl_easy_cleanup(ctxt.curl)
-    ctxt.curl = C_NULL
-end
-
-function process_response(ctxt::ConnContext, resp::Response)
-    resp_code = Int[1]
-    @ce_curl curl_easy_getinfo CURLINFO_RESPONSE_CODE resp_code
-
-    total_time = Float64[1.0]
-    @ce_curl curl_easy_getinfo CURLINFO_TOTAL_TIME total_time
-
-    resp.code = resp_code[1]
-    resp.total_time = total_time[1]
-end
-
-
 ##############################
 # Library initializations
 ##############################
@@ -302,383 +82,3 @@ ftp_init() = curl_global_init(CURL_GLOBAL_ALL)
 Cleanup global libcurl.
 """
 ftp_cleanup() = curl_global_cleanup()
-
-
-##############################
-# GET
-##############################
-
-"""
-    ftp_get(
-        options::RequestOptions,
-        file_name::AbstractString,
-        save_path::AbstractString="";
-        mode::FTP_MODE=binary_mode,
-        verbose::Union{Bool,IOStream}=false,
-    )
-
-Download a file with a non-persistent connection. Returns a `Response`.
-
-# Arguments
-* `options::RequestOptions`: the connection options. See `RequestOptions` for details.
-* `file_name::AbstractString`: the path to the file on the server.
-* `save_path::AbstractString=""`: if not specified the file is written to the `Response`
-    body.
-* `mode::FTP_MODE=binary_mode`: defines whether the file is transferred in binary or ASCII
-    format.
-* `verbose::Union{Bool,IOStream}=false`: an `IOStream` to capture LibCurl's output or a
-    `Bool`, if true output is written to STDERR.
-"""
-function ftp_get(
-    options::RequestOptions,
-    file_name::AbstractString,
-    save_path::AbstractString="";
-    mode::FTP_MODE=binary_mode,
-    verbose::Union{Bool,IOStream}=false,
-)
-    ctxt = setup_easy_handle(options)
-    try
-        return ftp_get(ctxt, file_name, save_path; mode=mode, verbose=verbose)
-    finally
-        cleanup_easy_context(ctxt)
-    end
-end
-
-"""
-    ftp_get(
-        ctxt::ConnContext,
-        file_name::AbstractString,
-        save_path::AbstractString="";
-        mode::FTP_MODE=binary_mode,
-        verbose::Union{Bool,IOStream}=false,
-    )
-
-Download a file with a persistent connection. Returns a `Response`.
-
-# Arguments
-* `ctxt::ConnContext`: encompases the connection options defined via ftp_connect. See
-    `RequestOptions` for details.
-* `file_name::AbstractString`: the path to the file on the server.
-* `save_path::AbstractString=""`: if not specified the file is written to the `Response`
-    body.
-* `mode::FTP_MODE=binary_mode`: defines whether the file is transferred in binary or
-    ASCII format.
-* `verbose::Union{Bool,IOStream}=false`: an `IOStream` to capture LibCurl's output or a
-    `Bool`, if true output is written to STDERR.
-"""
-function ftp_get(
-    ctxt::ConnContext,
-    file_name::AbstractString,
-    save_path::AbstractString="";
-    mode::FTP_MODE=binary_mode,
-    verbose::Union{Bool,IOStream}=false,
-)
-    wd = WriteData()
-
-    if !isempty(save_path)
-        wd.buffer = open(save_path, "w")
-    end
-
-    try
-        p_wd = pointer_from_objref(wd)
-
-        # Force active mode
-        #@ce_curl curl_easy_setopt CURLOPT_FTP_USE_EPSV 0
-        #@ce_curl curl_easy_setopt CURLOPT_FTP_USE_EPRT 0
-        #@ce_curl curl_easy_setopt CURLOPT_FTPPORT "-"
-
-        @ce_curl curl_easy_setopt CURLOPT_WRITEFUNCTION C_WRITE_FILE_CB[]
-        @ce_curl curl_easy_setopt CURLOPT_WRITEDATA p_wd
-
-        @ce_curl curl_easy_setopt CURLOPT_PROXY_TRANSFER_MODE Int64(1)
-
-        ftp_url = url(ctxt)
-        full_url = ftp_url * file_name
-        if mode == binary_mode
-            @ce_curl curl_easy_setopt CURLOPT_URL full_url * ";type=i"
-        elseif mode == ascii_mode
-            @ce_curl curl_easy_setopt CURLOPT_URL full_url * ";type=a"
-        end
-
-        resp = ftp_perform(ctxt, verbose)
-
-        if isopen(wd.buffer)
-            seekstart(wd.buffer)
-        end
-
-        @ce_curl curl_easy_setopt CURLOPT_URL ftp_url
-
-        # Set it back to default
-        @ce_curl curl_easy_setopt CURLOPT_PROXY_TRANSFER_MODE Int64(0)
-
-        resp.bytes_recd = wd.bytes_recd
-        resp.body = wd.buffer
-
-        return resp
-    finally
-        if !isempty(save_path)
-            close(wd.buffer)
-        end
-    end
-end
-
-##############################
-# PUT
-##############################
-
-"""
-    ftp_put(
-        options::RequestOptions,
-        file_name::AbstractString,
-        file::IO;
-        mode::FTP_MODE=binary_mode,
-        verbose::Union{Bool,IOStream}=false,
-    )
-
-Upload file with non-persistent connection. Returns a Response.
-
-# Arguments
-* `options::RequestOptions`: the connection options. See `RequestOptions` for details.
-* `file_name::AbstractString`: the path to the file on the server.
-* `file::IO`: what is being written to the server.
-* `mode::FTP_MODE=binary_mode`: defines whether the file is transferred in binary or
-    ASCII format.
-* `verbose::Union{Bool,IOStream}=false`: an `IOStream` to capture LibCurl's output or a
-    `Bool`, if true output is written to STDERR.
-"""
-function ftp_put(
-    options::RequestOptions,
-    file_name::AbstractString,
-    file::IO;
-    mode::FTP_MODE=binary_mode,
-    verbose::Union{Bool,IOStream}=false,
-)
-    ctxt = setup_easy_handle(options)
-    try
-        return ftp_put(ctxt, file_name, file; mode=mode, verbose=verbose)
-    finally
-        cleanup_easy_context(ctxt)
-    end
-end
-
-"""
-    ftp_put(
-        ctxt::ConnContext,
-        file_name::AbstractString,
-        file::IO;
-        mode::FTP_MODE=binary_mode,
-        verbose::Union{Bool,IOStream}=false,
-    )
-
-Upload file with persistent connection. Returns a `Response`.
-
-# Arguments
-* `ctxt::ConnContext`: encompases the connection options defined via ftp_connect. See
-    `RequestOptions` for details.
-* `file_name::AbstractString`: the path to the file on the server.
-* `file::IO`: what is being written to the server.
-* `mode::FTP_MODE=binary_mode`: defines whether the file is transferred in binary or
-    ASCII format.
-* `verbose::Union{Bool,IOStream}=false`: an `IOStream` to capture LibCurl's output or a
-    `Bool`, if true output is written to STDERR.
-"""
-function ftp_put(
-    ctxt::ConnContext,
-    file_name::AbstractString,
-    file::IO;
-    mode::FTP_MODE=binary_mode,
-    verbose::Union{Bool,IOStream}=false,
-)
-    rd = ReadData()
-
-    rd.src = file
-    seekend(file)
-    rd.sz = position(file)
-    seekstart(file)
-
-    p_rd = pointer_from_objref(rd)
-
-    @ce_curl curl_easy_setopt CURLOPT_UPLOAD Int64(1)
-    @ce_curl curl_easy_setopt CURLOPT_READDATA p_rd
-    @ce_curl curl_easy_setopt CURLOPT_READFUNCTION C_CURL_READ_CB[]
-
-    ftp_url = url(ctxt)
-    @ce_curl curl_easy_setopt CURLOPT_URL ftp_url * file_name
-
-    if mode == binary_mode
-        @ce_curl curl_easy_setopt CURLOPT_TRANSFERTEXT Int64(0)
-    elseif mode == ascii_mode
-        @ce_curl curl_easy_setopt CURLOPT_TRANSFERTEXT Int64(1)
-    end
-
-    @ce_curl curl_easy_setopt CURLOPT_INFILESIZE Int64(rd.sz)
-
-    resp = ftp_perform(ctxt, verbose)
-
-    # resest handle defaults
-    @ce_curl curl_easy_setopt CURLOPT_URL ftp_url
-    @ce_curl curl_easy_setopt CURLOPT_UPLOAD Int64(0)
-    @ce_curl curl_easy_setopt CURLOPT_TRANSFERTEXT Int64(0)
-
-    return resp
-end
-
-
-##############################
-# COMMAND
-##############################
-
-"""
-    ftp_command(
-        options::RequestOptions,
-        cmd::AbstractString;
-        verbose::Union{Bool,IOStream}=false,
-    )
-
-Pass FTP command with non-persistent connection. Returns a `Response`.
-"""
-function ftp_command(
-    options::RequestOptions,
-    cmd::AbstractString;
-    verbose::Union{Bool,IOStream}=false,
-)
-    ctxt = setup_easy_handle(options)
-    try
-        return ftp_command(ctxt, cmd; verbose=verbose)
-    finally
-        cleanup_easy_context(ctxt)
-    end
-end
-
-"""
-    ftp_command(
-        ctxt::ConnContext,
-        cmd::AbstractString;
-        verbose::Union{Bool,IOStream}=false,
-    )
-
-Pass FTP command with persistent connection. Returns a `Response`.
-"""
-function ftp_command(
-    ctxt::ConnContext,
-    cmd::AbstractString;
-    verbose::Union{Bool,IOStream}=false,
-)
-    wd = WriteData()
-    p_wd = pointer_from_objref(wd)
-
-    @ce_curl curl_easy_setopt CURLOPT_CUSTOMREQUEST cmd
-
-    @ce_curl curl_easy_setopt CURLOPT_WRITEFUNCTION C_WRITE_FILE_CB[]
-    @ce_curl curl_easy_setopt CURLOPT_WRITEDATA p_wd
-
-    resp = ftp_perform(ctxt, verbose)
-
-    resp.body = seekstart(wd.buffer)
-    resp.bytes_recd = wd.bytes_recd
-
-    parts = split(cmd, ' ', limit=2)
-    if resp.code == 250 && parts[1] == "CWD"
-        ctxt.url *= trailing(parts[2], '/')
-    end
-
-    return resp
-end
-
-
-##############################
-# CONNECT
-##############################
-
-"""
-    ftp_connect(options::RequestOptions; verbose::Union{Bool,IOStream}=false)
-
-Establish connection to FTP server. Returns a `ConnContext` and a `Response`.
-"""
-function ftp_connect(options::RequestOptions; verbose::Union{Bool,IOStream}=false)
-    ctxt = setup_easy_handle(options)
-    try
-        # ping the server
-        resp = ftp_command(ctxt, "LIST"; verbose=verbose)
-
-        return ctxt, resp
-    catch
-        cleanup_easy_context(ctxt)
-        rethrow()
-    end
-end
-
-
-##############################
-# CLOSE
-##############################
-
-"""
-    ftp_close_connection(ctxt::ConnContext)
-
-Close the connection to the FTP server.
-"""
-function ftp_close_connection(ctxt::ConnContext)
-    cleanup_easy_context(ctxt)
-    nothing
-end
-
-
-##############################
-# PERFORM
-##############################
-
-function ftp_perform(ctxt::ConnContext, verbose::Union{Bool,IOStream})
-    resp = Response()
-    p_resp = pointer_from_objref(resp)
-
-    @ce_curl curl_easy_setopt CURLOPT_HEADERFUNCTION C_HEADER_COMMAND_CB[]
-    @ce_curl curl_easy_setopt CURLOPT_HEADERDATA p_resp
-
-    libc_file = nothing
-    if verbose != false
-        @ce_curl curl_easy_setopt CURLOPT_VERBOSE Int64(1)
-        if isa(verbose, IOStream)
-
-            # flush the IOStream before making a duplicate Libc.FILE that will
-            # capture the verbose
-            flush(verbose)
-            libc_file = Libc.FILE(verbose)
-
-            @ce_curl curl_easy_setopt CURLOPT_STDERR libc_file.ptr
-        end
-    else
-        @ce_curl curl_easy_setopt CURLOPT_VERBOSE Int64(0)
-    end
-
-    try
-        @ce_curl curl_easy_perform
-        process_response(ctxt, resp)
-    finally
-        if isa(libc_file, Libc.FILE)
-            # Need to flush the Libc.FILE to make sure content is written, closing the
-            # Libc.FILE does not reliably do this. Then update the IOStream pointer to
-            # point to the same position so that data can be appended to the stream.
-            ccall(:fflush, Cvoid, (Ptr{Cvoid},), libc_file.ptr)
-            seek(verbose, position(libc_file))
-            close(libc_file)
-
-            @ce_curl curl_easy_setopt CURLOPT_STDERR Libc.FILE(RawFD(2), "w").ptr
-        end
-    end
-
-    return resp
-end
-
-##############################
-# ISEQUAL
-##############################
-
-function ==(this::RequestOptions, other::RequestOptions)
-    return (
-        this.uri == other.uri &&
-        this.ssl == other.ssl &&
-        this.verify_peer == other.verify_peer &&
-        this.active_mode == other.active_mode
-    )
-end
